refactor(midi): 重构MIDI时间计算系统以提高精度和性能
- 将Note结构体中的start/end浮点秒时间改为start_tick/end_tick整数tick值
- 修改tick_to_seconds函数参数类型从u32提升到u64以支持更大范围
- 更新音频引擎中所有基于tick的时间转换逻辑
- 移除全局持续时间跟踪改为按需从tick长度计算
- 优化音符排序算法从基于时间改为基于tick值
- 重构时间解析器避免不必要的浮点运算
- 更新播放控制和传输条中的时间戳处理逻辑

+++ crates/yinhe-egui/src/app.rs
@@ -368,7 +368,7 @@ impl eframe::App for App {
                     doc.playback.stop();
                 } else {
                     let tick = doc.cursor_tick.unwrap_or(0.0);
-                    let cursor_sample = (doc.midi.tick_to_seconds(tick as u32) * audio.sample_rate as f64) as u64;
+                    let cursor_sample = (doc.midi.tick_to_seconds(tick as u64) * audio.sample_rate as f64) as u64;
                     let engine_sample = handle.sample_position();
                     // If cursor is at the engine's position, just resume (no seek)
                     if cursor_sample.abs_diff(engine_sample) < (audio.sample_rate as u64 / 10) {
